@@ -2,10 +2,10 @@ package com.codingwithmitch.googlemaps2018.ui;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
+
 import android.util.Log;
 import android.view.View;
 import android.view.WindowManager;
@@ -17,6 +17,7 @@ import com.codingwithmitch.googlemaps2018.R;
 import com.codingwithmitch.googlemaps2018.models.User;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
+import com.google.android.material.snackbar.Snackbar;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.firestore.DocumentReference;
@@ -81,8 +82,9 @@ public class RegisterActivity extends AppCompatActivity implements
                             user.setUser_id(FirebaseAuth.getInstance().getUid());
 
                             FirebaseFirestoreSettings settings = new FirebaseFirestoreSettings.Builder()
-                                    .setTimestampsInSnapshotsEnabled(true)
+
                                     .build();
+                            //.setTimestampsInSnapshotsEnabled(true)
                             mDb.setFirestoreSettings(settings);
 
                             DocumentReference newUserRef = mDb
