@@ -1,65 +1,580 @@
 package com.codingwithmitch.googlemaps2018.ui;
 
+import android.Manifest;
+import android.animation.ObjectAnimator;
+import android.annotation.SuppressLint;
+import android.app.AlertDialog;
+import android.content.DialogInterface;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+
+
+import androidx.annotation.Nullable;
+import androidx.core.app.ActivityCompat;
+import androidx.core.content.ContextCompat;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
+
+import android.os.Handler;
+import android.os.Looper;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.RelativeLayout;
+import android.widget.Toast;
 
 
 import com.codingwithmitch.googlemaps2018.R;
 import com.codingwithmitch.googlemaps2018.adapters.UserRecyclerAdapter;
+import com.codingwithmitch.googlemaps2018.models.ClusterMarker;
 import com.codingwithmitch.googlemaps2018.models.User;
+import com.codingwithmitch.googlemaps2018.models.UserLocation;
+import com.codingwithmitch.googlemaps2018.util.MyClusterManagerRenderer;
+import com.codingwithmitch.googlemaps2018.util.ViewWeightAnimationWrapper;
+import com.google.android.gms.maps.CameraUpdateFactory;
+import com.google.android.gms.maps.GoogleMap;
+import com.google.android.gms.maps.MapView;
+import com.google.android.gms.maps.OnMapReadyCallback;
+import com.google.android.gms.maps.model.BitmapDescriptorFactory;
+import com.google.android.gms.maps.model.LatLng;
+import com.google.android.gms.maps.model.LatLngBounds;
+import com.google.android.gms.maps.model.Marker;
+import com.google.android.gms.maps.model.MarkerOptions;
+import com.google.android.gms.maps.model.Polyline;
+import com.google.android.gms.maps.model.PolylineOptions;
+import com.google.android.gms.tasks.OnCompleteListener;
+import com.google.android.gms.tasks.Task;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.firestore.DocumentReference;
+import com.google.firebase.firestore.DocumentSnapshot;
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.maps.DirectionsApiRequest;
+import com.google.maps.GaeRequestHandler;
+import com.google.maps.GeoApiContext;
+import com.google.maps.GeocodingApi;
+import com.google.maps.PendingResult;
+import com.google.maps.android.clustering.ClusterManager;
+import com.google.maps.errors.ApiException;
+import com.google.maps.internal.PolylineEncoding;
+import com.google.maps.model.DirectionsResult;
+import com.google.maps.model.DirectionsRoute;
+import com.google.maps.model.GeocodingResult;
 
+import java.io.IOException;
 import java.util.ArrayList;
+import java.util.List;
 
-public class UserListFragment extends Fragment {
+import static com.codingwithmitch.googlemaps2018.Constants.MAPVIEW_BUNDLE_KEY;
+
+public class UserListFragment extends Fragment implements
+        OnMapReadyCallback,
+        View.OnClickListener,
+        GoogleMap.OnInfoWindowClickListener,
+        GoogleMap.OnMarkerClickListener{
 
     private static final String TAG = "UserListFragment";
+    private static final int MAP_LAYOUT_STATE_CONTRACTED = 0;
+    private static final int MAP_LAYOUT_STATE_EXPANDED = 1;
+    private static final int LOCATION_UPDATE_INTERVAL = 3000;
 
     //widgets
     private RecyclerView mUserListRecyclerView;
-
+    private MapView mMapView;
+    private RelativeLayout mMapContainer;
+    private int mMapLayoutState = 0;
 
     //vars
     private ArrayList<User> mUserList = new ArrayList<>();
+    private ArrayList<UserLocation> mUserLocations = new ArrayList<>();
     private UserRecyclerAdapter mUserRecyclerAdapter;
 
+    //****** Realtime GPS update *****//
+    private Handler mHandler = new Handler();
+    private Runnable mRunnable;
+    //******** Google Maps ***** //
+    private GoogleMap mGoogleMap;
+    private LatLngBounds latLngBounds;
+    private UserLocation mUserPosition;
+    private GeoApiContext mGeoApiContext = null;
+    //************* Cluster Marker ********** //
+    private ClusterManager<ClusterMarker> mClusterManager;
+    private MyClusterManagerRenderer mClusterManagerRenderer;
+    private ArrayList<ClusterMarker> mClusterMarkers = new ArrayList<>();
 
-    public static UserListFragment newInstance(){
+    public static UserListFragment newInstance() {
         return new UserListFragment();
     }
 
     @Override
     public void onCreate(@Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        if(getArguments() != null){
+        // to attach arguments
+        if (getArguments() != null) {
+            // to catch userList
             mUserList = getArguments().getParcelableArrayList(getString(R.string.intent_user_list));
+            // to catch UserLocations
+            mUserLocations = getArguments().getParcelableArrayList(getString(R.string.intent_user_locations));
         }
     }
 
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        View view  = inflater.inflate(R.layout.fragment_user_list, container, false);
+        View view = inflater.inflate(R.layout.fragment_user_list, container, false);
+        // widgets
         mUserListRecyclerView = view.findViewById(R.id.user_list_recycler_view);
-
+        mMapView = view.findViewById(R.id.user_list_map);
+        mMapContainer = view.findViewById(R.id.map_container);
+        view.findViewById(R.id.btn_full_screen_map).setOnClickListener(this);
+        // initialize RecycleView
         initUserListRecyclerView();
+        // initialize GoogleMap
+        initGoogleMap(savedInstanceState);
+        setUserPostion();
+        // for debugging
+        for(UserLocation userLocation: mUserLocations){
+            Log.d(TAG, "onCreateView: user location: " + userLocation.getUser().getUsername());
+        }
         return view;
     }
 
+    //****** Realtime GPS update *****//
+    private void startUserLocationsRunnable(){
+        Log.d(TAG, "startUserLocationsRunnable: starting runnable for retrieving updated locations.");
+        mHandler.postDelayed(mRunnable = new Runnable() {
+            @Override
+            public void run() {
+                retrieveUserLocations();
+                mHandler.postDelayed(mRunnable, LOCATION_UPDATE_INTERVAL);
+            }
+        }, LOCATION_UPDATE_INTERVAL);
+    }
 
-    private void initUserListRecyclerView(){
+    private void stopLocationUpdates(){
+        mHandler.removeCallbacks(mRunnable);
+    }
+
+    private void retrieveUserLocations(){
+        Log.d(TAG, "retrieveUserLocations: retrieving location of all users in the chatroom.");
+
+        try{
+            for(final ClusterMarker clusterMarker: mClusterMarkers){
+
+                DocumentReference userLocationRef = FirebaseFirestore.getInstance()
+                        .collection(getString(R.string.collection_user_locations))
+                        .document(clusterMarker.getUser().getUser_id());
+
+                userLocationRef.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<DocumentSnapshot> task) {
+                        if(task.isSuccessful()){
+
+                            final UserLocation updatedUserLocation = task.getResult().toObject(UserLocation.class);
+
+                            // update the location
+                            for (int i = 0; i < mClusterMarkers.size(); i++) {
+                                try {
+                                    if (mClusterMarkers.get(i).getUser().getUser_id().equals(updatedUserLocation.getUser().getUser_id())) {
+
+                                        LatLng updatedLatLng = new LatLng(
+                                                updatedUserLocation.getGeo_point().getLatitude(),
+                                                updatedUserLocation.getGeo_point().getLongitude()
+                                        );
+
+                                        mClusterMarkers.get(i).setPosition(updatedLatLng);
+                                        // to update markers
+                                        mClusterManagerRenderer.setUpdateMarker(mClusterMarkers.get(i));
+                                    }
+
+
+                                } catch (NullPointerException e) {
+                                    Log.e(TAG, "retrieveUserLocations: NullPointerException: " + e.getMessage());
+                                }
+                            }
+                        }
+                    }
+                });
+            }
+        }catch (IllegalStateException e){
+            Log.e(TAG, "retrieveUserLocations: Fragment was destroyed during Firestore query. Ending query." + e.getMessage() );
+        }
+
+    }
+    //***** map marker ******//
+    private void addMapMarkers(){
+
+        if(mGoogleMap != null){
+
+            if(mClusterManager == null){
+                mClusterManager = new ClusterManager<ClusterMarker>(getActivity().getApplicationContext(), mGoogleMap);
+            }
+            if(mClusterManagerRenderer == null){
+                mClusterManagerRenderer = new MyClusterManagerRenderer(
+                        getActivity(),
+                        mGoogleMap,
+                        mClusterManager
+                );
+                mClusterManager.setRenderer(mClusterManagerRenderer);
+            }
+
+            for(UserLocation userLocation: mUserLocations){
+
+                Log.d(TAG, "addMapMarkers: location: " + userLocation.getGeo_point().toString());
+
+                try{
+                    String snippet;
+                    if(userLocation.getUser().getUser_id().equals(FirebaseAuth.getInstance().getUid())){
+                        snippet = "This is you";
+
+                    }
+                    else{
+                        snippet = "Determine route to " + userLocation.getUser().getUsername() + "?";
+                    }
+
+                    int avatar = R.drawable.cartman_cop; // set the default avatar
+                    try{
+                        avatar = Integer.parseInt(userLocation.getUser().getAvatar());
+                    }catch (NumberFormatException e){
+                        Log.d(TAG, "addMapMarkers: no avatar for " + userLocation.getUser().getUsername() + ", setting default.");
+                    }
+                    ClusterMarker newClusterMarker = new ClusterMarker(
+                            new LatLng(userLocation.getGeo_point().getLatitude(),userLocation.getGeo_point().getLongitude()),
+                            userLocation.getUser().getUsername(),
+                            snippet,
+                            avatar,
+                            userLocation.getUser()
+                    );
+                    mClusterManager.addItem(newClusterMarker);
+                    mClusterMarkers.add(newClusterMarker);
+
+                }catch (NullPointerException e){
+                    Log.e(TAG, "addMapMarkers: NullPointerException: " + e.getMessage() );
+                }
+
+            }
+            mClusterManager.cluster();
+
+            setCameraView();
+        }
+    }
+    //****************************//
+
+    private void setUserPostion(){
+        for (UserLocation userLocation:mUserLocations){
+            if (userLocation.getUser().getUser_id().equals(FirebaseAuth.getInstance().getUid())){
+                mUserPosition = userLocation;
+            }
+        }
+    }
+
+    private void setCameraView(){
+        // mapview window
+        double buttomBound = mUserPosition.getGeo_point().getLatitude() -.1;
+        double leftBound = mUserPosition.getGeo_point().getLongitude() -.1;
+        double topBound = mUserPosition.getGeo_point().getLatitude() +.1 ;
+        double rightBound = mUserPosition.getGeo_point().getLongitude() +.1 ;
+
+        latLngBounds = new LatLngBounds(
+                new LatLng(buttomBound,leftBound),
+                new LatLng(topBound,rightBound)
+        );
+
+        mGoogleMap.moveCamera(CameraUpdateFactory.newLatLngBounds(latLngBounds,0));
+    }
+
+    private void initGoogleMap(Bundle savedInstanceState){
+        // *** IMPORTANT ***
+        // MapView requires that the Bundle you pass contain _ONLY_ MapView SDK
+        // objects or sub-Bundles.
+        Bundle mapViewBundle = null;
+        if (savedInstanceState != null) {
+            mapViewBundle = savedInstanceState.getBundle(MAPVIEW_BUNDLE_KEY);
+        }
+
+        mMapView.onCreate(mapViewBundle);
+
+        mMapView.getMapAsync(this);
+
+        // api directions
+//        GeoApiContext context = new GeoApiContext.Builder()
+//                .apiKey(getString(R.string.google_map_api_key))
+//                .build();
+//        GeocodingResult[] results = new GeocodingResult[0];
+//        try {
+//            results = GeocodingApi.geocode(context,
+//                    "1600 Amphitheatre Parkway Mountain View, CA 94043").await();
+//        } catch (ApiException e) {
+//            e.printStackTrace();
+//        } catch (InterruptedException e) {
+//            e.printStackTrace();
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//        Gson gson = new GsonBuilder().setPrettyPrinting().create();
+//        System.out.println(gson.toJson(results[0].addressComponents));
+//        Toast.makeText(getContext(),results[0].addressComponents.toString(),Toast.LENGTH_LONG).show();
+            if (mGeoApiContext == null) {
+                mGeoApiContext = new GeoApiContext.Builder()
+                        .apiKey(getString(R.string.google_map_api_key))
+                        .build();
+            }
+
+    }
+
+    private void initUserListRecyclerView() {
         mUserRecyclerAdapter = new UserRecyclerAdapter(mUserList);
         mUserListRecyclerView.setAdapter(mUserRecyclerAdapter);
         mUserListRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
     }
+
+    @Override
+    public void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+
+        Bundle mapViewBundle = outState.getBundle(MAPVIEW_BUNDLE_KEY);
+        if (mapViewBundle == null) {
+            mapViewBundle = new Bundle();
+            outState.putBundle(MAPVIEW_BUNDLE_KEY, mapViewBundle);
+        }
+
+        mMapView.onSaveInstanceState(mapViewBundle);
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
+        mMapView.onResume();
+        // to realtime update gps
+        startUserLocationsRunnable();
+    }
+
+    @Override
+    public void onStart() {
+        super.onStart();
+        mMapView.onStart();
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        mMapView.onStop();
+    }
+
+    @SuppressLint("PotentialBehaviorOverride")
+    @Override
+    public void onMapReady(GoogleMap map) {
+        if (ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_FINE_LOCATION)
+                != PackageManager.PERMISSION_GRANTED
+                && ActivityCompat.checkSelfPermission(getActivity(), Manifest.permission.ACCESS_COARSE_LOCATION)
+                != PackageManager.PERMISSION_GRANTED) {
+            return;
+        }
+        mGoogleMap = map;
+        map.setMyLocationEnabled(true);
+        markerTest(mGoogleMap);
+        setCameraView();
+//        addMapMarkers();
+
+        mGoogleMap.setOnInfoWindowClickListener(this);
+        mGoogleMap.setOnMarkerClickListener(this);
+    }
+
+    private void markerTest(GoogleMap mGoogleMap) {
+        LatLng coordinate = new LatLng(31.0671364,31.404148);
+        Marker m =   mGoogleMap.addMarker(new MarkerOptions()
+                .position(coordinate).title("إختبار")
+                .icon(BitmapDescriptorFactory.defaultMarker(BitmapDescriptorFactory.HUE_GREEN))) ;
+        m.setTag(0);
+        LatLng latLng = new LatLng(m.getPosition().latitude,m.getPosition().longitude);
+        mGoogleMap.addMarker(new MarkerOptions().position(latLng));
+    }
+
+    @Override
+    public void onPause() {
+        mMapView.onPause();
+        super.onPause();
+    }
+
+    @Override
+    public void onDestroy() {
+        mMapView.onDestroy();
+        super.onDestroy();
+        // to stop realtime update
+        stopLocationUpdates();
+    }
+
+    @Override
+    public void onLowMemory() {
+        super.onLowMemory();
+        mMapView.onLowMemory();
+    }
+
+//******* expanding and contracting ********** //
+private void expandMapAnimation(){
+    ViewWeightAnimationWrapper mapAnimationWrapper = new ViewWeightAnimationWrapper(mMapContainer);
+    ObjectAnimator mapAnimation = ObjectAnimator.ofFloat(mapAnimationWrapper,
+            "weight",
+            50,
+            100);
+    mapAnimation.setDuration(800);
+
+    ViewWeightAnimationWrapper recyclerAnimationWrapper = new ViewWeightAnimationWrapper(mUserListRecyclerView);
+    ObjectAnimator recyclerAnimation = ObjectAnimator.ofFloat(recyclerAnimationWrapper,
+            "weight",
+            50,
+            0);
+    recyclerAnimation.setDuration(800);
+
+    recyclerAnimation.start();
+    mapAnimation.start();
+}
+    private void contractMapAnimation(){
+        ViewWeightAnimationWrapper mapAnimationWrapper = new ViewWeightAnimationWrapper(mMapContainer);
+        ObjectAnimator mapAnimation = ObjectAnimator.ofFloat(mapAnimationWrapper,
+                "weight",
+                100,
+                50);
+        mapAnimation.setDuration(800);
+
+        ViewWeightAnimationWrapper recyclerAnimationWrapper = new ViewWeightAnimationWrapper(mUserListRecyclerView);
+        ObjectAnimator recyclerAnimation = ObjectAnimator.ofFloat(recyclerAnimationWrapper,
+                "weight",
+                0,
+                50);
+        recyclerAnimation.setDuration(800);
+
+        recyclerAnimation.start();
+        mapAnimation.start();
+    }
+
+    @Override
+    public void onClick(View v) {
+        switch (v.getId()){
+            case R.id.btn_full_screen_map:{
+
+                if(mMapLayoutState == MAP_LAYOUT_STATE_CONTRACTED){
+                    mMapLayoutState = MAP_LAYOUT_STATE_EXPANDED;
+                    expandMapAnimation();
+                }
+                else if(mMapLayoutState == MAP_LAYOUT_STATE_EXPANDED){
+                    mMapLayoutState = MAP_LAYOUT_STATE_CONTRACTED;
+                    contractMapAnimation();
+                }
+                break;
+            }
+
+        }
+    }
+
+    @Override
+    public void onInfoWindowClick(final Marker marker) {
+//        if(marker.getSnippet().equals("This is you")){
+//            marker.hideInfoWindow();
+//        }
+//        else{
+
+            final AlertDialog.Builder builder = new AlertDialog.Builder(getActivity());
+            builder.setMessage("marker.getSnippet()")
+                    .setCancelable(true)
+                    .setPositiveButton("Yes", new DialogInterface.OnClickListener() {
+                        public void onClick(@SuppressWarnings("unused") final DialogInterface dialog, @SuppressWarnings("unused") final int id) {
+                            calculateDirections(marker);
+                            dialog.dismiss();
+                        }
+                    })
+                    .setNegativeButton("No", new DialogInterface.OnClickListener() {
+                        public void onClick(final DialogInterface dialog, @SuppressWarnings("unused") final int id) {
+                            dialog.cancel();
+                        }
+                    });
+            final AlertDialog alert = builder.create();
+            alert.show();
+//        }
+    }
+//******************* Calculating Directions ************************
+    private void calculateDirections(Marker marker){
+    Log.d(TAG, "calculateDirections: calculating directions.");
+
+    com.google.maps.model.LatLng destination = new com.google.maps.model.LatLng(
+            marker.getPosition().latitude,
+            marker.getPosition().longitude
+    );
+    DirectionsApiRequest directions = new DirectionsApiRequest(mGeoApiContext);
+
+    directions.alternatives(true);
+    directions.origin(
+            new com.google.maps.model.LatLng(
+                    mUserPosition.getGeo_point().getLatitude(),
+                    mUserPosition.getGeo_point().getLongitude()
+            )
+    );
+    Log.d(TAG, "calculateDirections: destination: " + destination.toString());
+    directions.destination(destination).setCallback(new PendingResult.Callback<DirectionsResult>() {
+        @Override
+        public void onResult(DirectionsResult result) {
+                Log.d(TAG, "calculateDirections: routes: " + result.routes[0].toString());
+                Log.d(TAG, "calculateDirections: duration: " + result.routes[0].legs[0].duration);
+                Log.d(TAG, "calculateDirections: distance: " + result.routes[0].legs[0].distance);
+                Log.d(TAG, "calculateDirections: geocodedWayPoints: " + result.geocodedWaypoints[0].toString());
+
+            Log.d(TAG, "onResult: successfully retrieved directions.");
+            addPolylinesToMap(result);
+        }
+
+        @Override
+        public void onFailure(Throwable e) {
+            Log.e(TAG, "calculateDirections: Failed to get directions: " + e.getMessage() );
+
+        }
+    });
 }
 
+
+    @Override
+    public boolean onMarkerClick(Marker marker) {
+        Toast.makeText(getContext(),"soe",Toast.LENGTH_LONG).show();
+        return false;
+    }
+
+    private void addPolylinesToMap(final DirectionsResult result){
+        new Handler(Looper.getMainLooper()).post(new Runnable() {
+            @Override
+            public void run() {
+                Log.d(TAG, "run: result routes: " + result.routes.length);
+
+                for(DirectionsRoute route: result.routes){
+                    Log.d(TAG, "run: leg: " + route.legs[0].toString());
+                    List<com.google.maps.model.LatLng> decodedPath = PolylineEncoding.decode(route.overviewPolyline.getEncodedPath());
+
+                    List<LatLng> newDecodedPath = new ArrayList<>();
+
+                    // This loops through all the LatLng coordinates of ONE polyline.
+                    for(com.google.maps.model.LatLng latLng: decodedPath){
+
+//                        Log.d(TAG, "run: latlng: " + latLng.toString());
+
+                        newDecodedPath.add(new LatLng(
+                                latLng.lat,
+                                latLng.lng
+                        ));
+                    }
+                    Polyline polyline = mGoogleMap.addPolyline(new PolylineOptions().addAll(newDecodedPath));
+                    polyline.setColor(ContextCompat.getColor(getActivity(), R.color.darkGrey));
+                    polyline.setClickable(true);
+
+                }
+            }
+        });
+    }
+
+
+
+}
 
 
 
