@@ -1,6 +1,8 @@
 package com.codingwithmitch.googlemaps2018;
 
 public class Constants {
-
+    public static final String MAPVIEW_BUNDLE_KEY = "MapViewBundleKey";
     public static final int ERROR_DIALOG_REQUEST = 9001;
+    public static final int PERMISSIONS_REQUEST_ENABLE_GPS = 9002;
+    public static final int PERMISSIONS_REQUEST_ACCESS_FINE_LOCATION = 9003;
 }
